# backend/storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'E-commerce Modular API'
  description: 'Tài liệu API hệ thống E-commerce (Sử dụng PHP 8 Attributes)'
  contact:
    email: admin@system.com
  version: 1.0.0
servers:
  -
    url: /api
    description: 'API Server'
paths:
...
  /payments:
    get:
      tags:
        - Payments
      summary: 'Lịch sử giao dịch (User xem của mình, Admin xem tất cả)'
      operationId: d2cb9e33536e2a3d10ea0e21861315fd
      parameters:
        -
          name: page
          in: query
          schema:
            type: integer
        -
          name: status
          in: query
          schema:
            type: string
            enum:
              - pending
              - completed
              - failed
      responses:
        200:
          description: Success
      security:
        -
          bearerAuth: []
    post:
      tags:
        - Payments
      summary: 'Tạo yêu cầu thanh toán (Thường dùng khi Retry payment)'
      operationId: 919d9707a0f0b62ff9bb94318680424b
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - order_uuid
                - payment_method
              properties:
                order_uuid:
                  type: string
                  format: uuid
                payment_method:
                  type: string
                  enum:
                    - cod
                    - banking
                    - momo
                    - stripe
              type: object
      responses:
        200:
          description: 'Payment Initiated (Return payment URL)'
      security:
        -
          bearerAuth: []
  '/payments/{uuid}':
    get:
      tags:
        - Payments
      summary: 'Xem chi tiết giao dịch'
      operationId: 2a56147aab532093b44a043f87eaa1d9
      parameters:
        -
          name: uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Success
      security:
        -
          bearerAuth: []
    put:
      tags:
        - Payments
      summary: 'Cập nhật trạng thái thanh toán (Admin Only - Manual)'
      operationId: 3e17e5a5f80fdbfe49a6536073361ca6
      requestBody:
        content:
          application/json:
            schema:
              properties:
                status:
                  type: string
                  enum:
                    - paid
                    - failed
                    - refunded
                transaction_id:
                  type: string
              type: object
      responses:
        200:
          description: Updated
      security:
        -
          bearerAuth: []
  '/payments/callback/{provider}':
    post:
      tags:
        - Payments
      summary: 'Webhook nhận kết quả thanh toán (Public)'
      description: 'API này dành cho Payment Gateway (Momo, Stripe, VNPAY) gọi lại. Không cần Token.'
      operationId: 62edbc63f5e9271506b563cdbf3d58b6
      parameters:
        -
          name: provider
          in: path
          required: true
          schema:
            type: string
            enum:
              - momo
              - vnpay
              - stripe
      responses:
        200:
          description: 'IPN Received'
  /public/products:
    get:
      tags:
        - Products
      summary: 'Lấy danh sách sản phẩm (Public - Chỉ hiện Active)'
      operationId: f089d65f31dc82c6e3b80e06c542463f
      parameters:
        -
          name: page
          in: query
          schema:
            type: integer
        -
          name: per_page
          in: query
          schema:
            type: integer
        -
          name: category_uuid
          in: query
          schema:
            type: string
        -
          name: search
          in: query
          schema:
            type: string
      responses:
        200:
          description: Success
  '/public/products/{uuid}':
    get:
      tags:
        - Products
      summary: 'Xem chi tiết sản phẩm (Public)'
      operationId: b96e795a251da1d56654b459cec75057
      parameters:
        -
          name: uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Success
        404:
          description: 'Not found'
  /admin/products:
    get:
      tags:
        - Products
      summary: 'Danh sách sản phẩm (Admin - Xem cả ẩn)'
      operationId: 90b09429681f249365da59a28870cfb5
      responses:
        200:
          description: Success
      security:
        -
          bearerAuth: []
    post:
      tags:
        - Products
      summary: 'Tạo sản phẩm mới (Admin)'
      operationId: 6cac846a61b72104adff1ee91bd8f69b
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - name
                - sku
                - price
                - category_uuid
              properties:
                name:
                  type: string
                category_uuid:
                  type: string
                  format: uuid
                price:
                  type: number
                sku:
                  type: string
              type: object
      responses:
        201:
          description: Created
      security:
        -
          bearerAuth: []
  '/admin/products/{uuid}':
    put:
      tags:
        - Products
      summary: 'Cập nhật sản phẩm (Admin)'
      operationId: ff7d7b298b0d62cf0b4503308d3af809
      parameters:
        -
          name: uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              properties:
                name:
                  type: string
                is_active:
                  type: boolean
              type: object
      responses:
        200:
          description: Updated
      security:
        -
          bearerAuth: []
    delete:
      tags:
        - Products
      summary: 'Xóa sản phẩm (Admin)'
      operationId: 997ef24991477503409e3edfd16a1553
      parameters:
        -
          name: uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Deleted
      security:
        -
          bearerAuth: []
  '/admin/products/{uuid}/images':
    post:
      tags:
        - 'Product Images'
      summary: 'Upload ảnh cho sản phẩm'
      operationId: 34871052eb1f635905f8fcb6c557c603
      parameters:
        -
          name: uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          multipart/form-data:
            schema:
              properties:
                image:
                  type: string
                  format: binary
                is_primary:
                  type: boolean
              type: object
      responses:
        201:
          description: Uploaded
      security:
        -
          bearerAuth: []
  '/admin/product-images/{uuid}':
    delete:
      tags:
        - 'Product Images'
      summary: 'Xóa ảnh sản phẩm'
      operationId: f7978e2e0e8a4fb631fc0685898fe068
      parameters:
        -
          name: uuid
          in: path
          description: 'UUID của Image'
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Deleted
      security:
        -
          bearerAuth: []
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  -
    bearerAuth: []